add unit test and small refactoring
